==> notebooks/04-model2-nn-training.ipynb <==
# 04 - model2 - neural net

### 1. Set up

import numpy as np
import pandas as pd
import tensorflow as tf
import os
from tensorflow import keras
from tensorflow.keras.layers.experimental.preprocessing import TextVectorization
from tensorflow.keras.layers import Embedding
from tensorflow.keras import layers, callbacks
from sklearn.metrics import classification_report
import pickle
import time

df = pd.read_csv('/home/jupyter/sb-entity-classification/data/data.csv')
df.columns = ['class','name']
df['class'] = df['class'] -1  # based on information provided in the brief

classes_list = pd.read_csv('/home/jupyter/sb-entity-classification/data/classes.txt', header = None)
classes_list['class'] = classes_list.index
classes_list.columns = ['class_name', 'class']
class_names = classes_list['class_name'].tolist()
df = df.merge(classes_list, on = 'class', how = 'left')

### 2. Split train, validation, and test set

# separate train validation and test
msk = np.random.rand(len(df)) < 0.98
train = df[msk]
test = df[~msk]

test.to_pickle('/home/jupyter/sb-entity-classification/data/test_nn.pkl')

train.shape

test.shape

samples = train['name'].tolist()
labels  = train['class'].tolist()

# Shuffle the data
seed = 42
rng = np.random.RandomState(seed)
rng.shuffle(samples)
rng = np.random.RandomState(seed)
rng.shuffle(labels)

# Extract a training & validation split
validation_split = 0.2
num_validation_samples = int(validation_split * len(samples))
train_samples = samples[:-num_validation_samples]
val_samples = samples[-num_validation_samples:]
train_labels = labels[:-num_validation_samples]
val_labels = labels[-num_validation_samples:]

### 3. Use pre-trained GloVe embeddings

vectorizer = TextVectorization(max_tokens=20000, output_sequence_length=8)
text_ds = tf.data.Dataset.from_tensor_slices(train_samples).batch(128)
vectorizer.adapt(text_ds)

voc = vectorizer.get_vocabulary()
word_index = dict(zip(voc, range(len(voc))))

path_to_glove_file = os.path.join(
    os.path.expanduser("~"), "sb-entity-classification/data/glove.6B/glove.6B.100d.txt"
)

embeddings_index = {}
with open(path_to_glove_file) as f:
    for line in f:
        word, coefs = line.split(maxsplit=1)
        coefs = np.fromstring(coefs, "f", sep=" ")
        embeddings_index[word] = coefs

print("Found %s word vectors." % len(embeddings_index))

num_tokens = len(voc) + 2
embedding_dim = 100
hits = 0
misses = 0

# Prepare embedding matrix
embedding_matrix = np.zeros((num_tokens, embedding_dim))
for word, i in word_index.items():
    embedding_vector = embeddings_index.get(word)
    if embedding_vector is not None:
        # Words not found in embedding index will be all-zeros.
        # This includes the representation for "padding" and "OOV"
        embedding_matrix[i] = embedding_vector
        hits += 1
    else:
        misses += 1
print("Converted %d words (%d misses)" % (hits, misses))

x_train = vectorizer(np.array([[s] for s in train_samples])).numpy()
x_val = vectorizer(np.array([[s] for s in val_samples])).numpy()

y_train = np.array(train_labels)
y_val = np.array(val_labels)

### 4. Build NN Model

embedding_layer = Embedding(
    num_tokens,
    embedding_dim,
    embeddings_initializer=keras.initializers.Constant(embedding_matrix),
    trainable=False,
)

int_sequences_input = keras.Input(shape=(None,), dtype="int64")
embedded_sequences = embedding_layer(int_sequences_input)
x = layers.Conv1D(128, 1, activation="relu")(embedded_sequences)
x = layers.MaxPooling1D(1)(x)
x = layers.Conv1D(128, 1, activation="relu")(x)
x = layers.MaxPooling1D(1)(x)
x = layers.Conv1D(128, 1, activation="relu")(x)
x = layers.GlobalMaxPooling1D()(x)
x = layers.Dense(128, activation="relu")(x)
x = layers.Dropout(0.5)(x)
preds = layers.Dense(len(class_names), activation="softmax")(x)
model = keras.Model(int_sequences_input, preds)
model.summary()

early_stopping_cb = callbacks.EarlyStopping(
    monitor='val_acc',
    min_delta=0,
    patience=2,
    verbose=0,
    mode='auto',
    baseline=None,
    restore_best_weights=True)

cbs = [
    early_stopping_cb
]

### 5. Train Model

model.compile(loss="sparse_categorical_crossentropy", 
              optimizer="rmsprop", 
              metrics=["acc"])
model.fit(x_train, 
          y_train, 
          batch_size=128, 
          epochs=20, 
          validation_data=(x_val, y_val),
          shuffle=True,
          callbacks=cbs)

string_input = keras.Input(shape=(1,), dtype="string")
x = vectorizer(string_input)
preds = model(x)
end_to_end_model = keras.Model(string_input, preds)

### save models and artifacts
# tf.keras.models.save_model(model, '/home/jupyter/sb-entity-classification/models/nn_embedding.h5')
# tf.keras.models.save_model(end_to_end_model, '/home/jupyter/sb-entity-classification/models/end_to_end_nn_embedding')
# with open('/home/jupyter/sb-entity-classification/models/class_names.txt', "wb") as fp:   
#     pickle.dump(class_names, fp)

### 6. Predict using trained model

def get_prediction(string_input):
    probabilities = end_to_end_model.predict([[string_input]])
    return class_names[np.argmax(probabilities[0])]

#### 6.1 try it on a few examples

test.sample(5)

get_prediction("Diplosphaera")

get_prediction("Anthony Ashley-Cooper (cricketer)")

get_prediction("Jeep Grand Cherokee")

get_prediction("Wombling Free")

get_prediction("Ulmus 'Folia Variegata Pendula'")

_Observations_:

From the sample examples, the model is not able to differentiate titles for different media of artworks (when a keyword is absent), which is probably only going to be able to resolved or improved with additional data sources of contexts. 

#### 6.2 try it on a sample of test set to estimate run time

sample_size = 100
test_sample = test.sample(sample_size)
start = time.time()
predictions = list(map(get_prediction,test_sample['name'].tolist()))
time_taken = time.time() - start
avg_time_p = time_taken / sample_size
print('Avg time per prediction {:.2f}s'.format(avg_time_p))
print('Total time for all {} forecasts: {:.2f}h'.format(sample_size, time_taken / 3600))

print(classification_report(test_sample['class_name'].tolist()[:sample_size], predictions, digits=3))

#### 6.3 to score on the entire test set:
Please see scripts/score_with_nn.py, which was developed using notebooks/05-model2-nn-scoring.ipynb
